pyproject.toml
[project]
name = "lego"
version = "0.1.0"
description = "Build scalable, maintainable, performant webapps one block at a time"
authors = [{ name = "Karthik Rajgopal", email = "karthik.rajgopal@hotmail.com" }]
requires-python = ">=3.12,<4.0"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "python-fasthtml (>=0.12.5,<0.13.0)",
    "monsterui (>=1.0.11,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "resend (>=2.7.0,<3.0.0)",
    "apscheduler>=3.11.0",
    "gunicorn>=23.0.0",
    "rcssmin>=1.2.1",
    "rjsmin>=1.2.4",
    "redis[hiredis]>=6.2.0",
    "ujson>=5.10.0",
    "fastsql>=2.0.3",
    "rclone-python>=0.1.23",
]

[project.urls]
Homepage = "https://github.com/Karthik777/lego"
Issues = "https://github.com/Karthik777/lego/issues"


[tool.hatch.build.targets.wheel]
packages = ["lego"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
